requirements.txt
json_repair~=0.59
numpy~=2.4
opencv-python~=4.13
PyJWT~=2.12
PyMuPDF~=1.27
qreader~=3.16
